Cleanup system lua and break out settings,zone,entry info
